requirements.txt
# Standard PyPI packages
diffusers
opencv-python
matplotlib
pypose
ipykernel
scipy
numpy
pillow
transformers
jupyter
ipywidgets
accelerate

# Packages from a different index
--find-links https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
